Move settings effects into a hook
